# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Restaurant API",
        "contact": {
            "name": "Hojune Yoo",
            "email": "soy0803@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "GetAllUser",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Save a user",
                "operationId": "SaveUser",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "User name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a User",
                "operationId": "GetUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "operationId": "UpdateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "User name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "operationId": "DeleteUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/visits": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Get all visits",
                "operationId": "GetAllVisits",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Visits"
                ],
                "summary": "Save a visit",
                "operationId": "SaveVisit",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "Visitor's user id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "restaurant_id",
                        "in": "formData",
                        "description": "Visited restaurant id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/visits/{visitId}": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Get a visit",
                "operationId": "GetVisit",
                "parameters": [
                    {
                        "name": "visitId",
                        "in": "path",
                        "description": "Visit id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Visits"
                ],
                "summary": "End visit",
                "operationId": "EndVisit",
                "parameters": [
                    {
                        "name": "visitId",
                        "in": "path",
                        "description": "Visit Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Visits"
                ],
                "summary": "Delete a visit",
                "operationId": "DeleteVisit",
                "parameters": [
                    {
                        "name": "visitId",
                        "in": "path",
                        "description": "Visit id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/visits/orers/{visitId}": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Get orders for this visit",
                "operationId": "GetVisitOrders",
                "parameters": [
                    {
                        "name": "visitId",
                        "in": "path",
                        "description": "Visit id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
...
}